Refactor: Separate Views and Controllers (#191)
