+++ mobile/lib/view/auth/ViewVerifyOTP.dart
@@ -1,10 +1,9 @@
 import 'package:bomcurriculo/include/BodyAuth.dart';
 import 'package:bomcurriculo/util/Translation.dart';
-import 'package:bomcurriculo/view/auth/ViewResetPassword.dart';
 import 'package:bomcurriculo/widget/WidgetError.dart';
 import 'package:flutter/material.dart';
 
-import '../../service/API.dart';
+import '../../controller/auth/ControllerVerifyOTP.dart';
 import '../../widget/WidgetButton.dart';
 import '../../widget/WidgetInputText.dart';
 
@@ -15,97 +14,21 @@ class ViewVerifyOTP extends StatefulWidget {
 }
 
 class _ViewVerifyOTP extends State<ViewVerifyOTP> {
-  String errorText = '';
-
-  final controllerOTP1 = TextEditingController();
-  final controllerOTP2 = TextEditingController();
-  final controllerOTP3 = TextEditingController();
-  final controllerOTP4 = TextEditingController();
-  final controllerOTP5 = TextEditingController();
-  final controllerOTP6 = TextEditingController();
-
-  final focusOTP1 = FocusNode();
-  final focusOTP2 = FocusNode();
-  final focusOTP3 = FocusNode();
-  final focusOTP4 = FocusNode();
-  final focusOTP5 = FocusNode();
-  final focusOTP6 = FocusNode();
-
-  void getTranslation() async {
-    await Translation.instance.load("pt-BR");
-    setState(() {});
-  }
+  late final ControllerVerifyOTP controller;
 
   @override
   void initState() {
     super.initState();
-    getTranslation();
-    focusOTP1.requestFocus();
+    controller = ControllerVerifyOTP(() {
+      if (mounted) setState(() {});
+    });
+    controller.init();
   }
 
   @override
   void dispose() {
-    controllerOTP1.dispose();
-    controllerOTP2.dispose();
-    controllerOTP3.dispose();
-    controllerOTP4.dispose();
-    controllerOTP5.dispose();
-    controllerOTP6.dispose();
-
-    focusOTP1.dispose();
-    focusOTP2.dispose();
-    focusOTP3.dispose();
-    focusOTP4.dispose();
-    focusOTP5.dispose();
-    focusOTP6.dispose();
-
+    controller.dispose();
     super.dispose();
-  }
-
-  void doConfirmOTP() async {
-    setState(() {
-      errorText = '';
-    });
-
-    if (controllerOTP1.text == "" ||
-        controllerOTP2.text == "" ||
-        controllerOTP3.text == "" ||
-        controllerOTP4.text == "" ||
-        controllerOTP5.text == "" ||
-        controllerOTP6.text == "") {
-      //erro
-      setState(() {
-        errorText = Translation.instance.translate('Invalid OTP');
-      });
-      return;
-    }
-
-    String otp = "";
-    otp += controllerOTP1.text;
-    otp += controllerOTP2.text;
-    otp += controllerOTP3.text;
-    otp += controllerOTP4.text;
-    otp += controllerOTP5.text;
-    otp += controllerOTP6.text;
-
-    API api = API();
-    var response = await api.post('auth/verify-otp', {'otp': otp});
-
-    if (!mounted) return;
-    if (response.statusCode == 200) {
-      Navigator.push(
-        context,
-        MaterialPageRoute(builder: (context) => ViewResetPassword(otp: otp)),
-      );
-    } else if (response.statusCode == 422) {
-      setState(() {
-        errorText = Translation.instance.translate('Invalid OTP');
-      });
-    } else {
-      setState(() {
-        errorText = Translation.instance.translate('Server comunication error');
-      });
-    }
   }
 
   @override
@@ -124,9 +47,9 @@ class _ViewVerifyOTP extends State<ViewVerifyOTP> {
             children: [
               Expanded(
                 child: WidgetInputText(
-                  controller: controllerOTP1,
-                  focusNode: focusOTP1,
-                  nextFocusNode: focusOTP2,
+                  controller: controller.controllerOTP1,
+                  focusNode: controller.focusOTP1,
+                  nextFocusNode: controller.focusOTP2,
                   maxLength: 1,
                   textAlignCenter: true,
                 ),
@@ -134,11 +57,11 @@ class _ViewVerifyOTP extends State<ViewVerifyOTP> {
               SizedBox(width: 5.0),
               Expanded(
                 child: WidgetInputText(
-                  controller: controllerOTP2,
-                  previousController: controllerOTP1,
-                  focusNode: focusOTP2,
-                  previousFocusNode: focusOTP1,
-                  nextFocusNode: focusOTP3,
+                  controller: controller.controllerOTP2,
+                  previousController: controller.controllerOTP1,
+                  focusNode: controller.focusOTP2,
+                  previousFocusNode: controller.focusOTP1,
+                  nextFocusNode: controller.focusOTP3,
                   maxLength: 1,
                   textAlignCenter: true,
                 ),
@@ -146,11 +69,11 @@ class _ViewVerifyOTP extends State<ViewVerifyOTP> {
               SizedBox(width: 5.0),
               Expanded(
                 child: WidgetInputText(
-                  controller: controllerOTP3,
-                  previousController: controllerOTP2,
-                  focusNode: focusOTP3,
-                  previousFocusNode: focusOTP2,
-                  nextFocusNode: focusOTP4,
+                  controller: controller.controllerOTP3,
+                  previousController: controller.controllerOTP2,
+                  focusNode: controller.focusOTP3,
+                  previousFocusNode: controller.focusOTP2,
+                  nextFocusNode: controller.focusOTP4,
                   maxLength: 1,
                   textAlignCenter: true,
                 ),
@@ -158,11 +81,11 @@ class _ViewVerifyOTP extends State<ViewVerifyOTP> {
               SizedBox(width: 5.0),
               Expanded(
                 child: WidgetInputText(
-                  controller: controllerOTP4,
-                  previousController: controllerOTP3,
-                  focusNode: focusOTP4,
-                  previousFocusNode: focusOTP3,
-                  nextFocusNode: focusOTP5,
+                  controller: controller.controllerOTP4,
+                  previousController: controller.controllerOTP3,
+                  focusNode: controller.focusOTP4,
+                  previousFocusNode: controller.focusOTP3,
+                  nextFocusNode: controller.focusOTP5,
                   maxLength: 1,
                   textAlignCenter: true,
                 ),
@@ -170,11 +93,11 @@ class _ViewVerifyOTP extends State<ViewVerifyOTP> {
               SizedBox(width: 5.0),
               Expanded(
                 child: WidgetInputText(
-                  controller: controllerOTP5,
-                  previousController: controllerOTP4,
-                  focusNode: focusOTP5,
-                  previousFocusNode: focusOTP4,
-                  nextFocusNode: focusOTP6,
+                  controller: controller.controllerOTP5,
+                  previousController: controller.controllerOTP4,
+                  focusNode: controller.focusOTP5,
+                  previousFocusNode: controller.focusOTP4,
+                  nextFocusNode: controller.focusOTP6,
                   maxLength: 1,
                   textAlignCenter: true,
                 ),
@@ -182,19 +105,19 @@ class _ViewVerifyOTP extends State<ViewVerifyOTP> {
               SizedBox(width: 5.0),
               Expanded(
                 child: WidgetInputText(
-                  controller: controllerOTP6,
-                  previousController: controllerOTP5,
-                  focusNode: focusOTP6,
-                  previousFocusNode: focusOTP5,
+                  controller: controller.controllerOTP6,
+                  previousController: controller.controllerOTP5,
+                  focusNode: controller.focusOTP6,
+                  previousFocusNode: controller.focusOTP5,
                   maxLength: 1,
                   textAlignCenter: true,
                 ),
               ),
             ],
           ),
-          WidgetError(text: errorText),
+          WidgetError(text: controller.errorText),
           GestureDetector(
-            onTap: doConfirmOTP,
+            onTap: () => controller.doConfirmOTP(context),
             child: WidgetButton(
               title: Translation.instance.translate('Confirm OTP'),
             ),
